# Minor pcolormesh optimization