The first few commits simply remove a few extra variables that duplicate other stuff.

The second-last commit changes the way that wrapped quads are determined by looking at the length of the edges instead of the angles. This limits the masked quads to _just_ the ones that wrap, whereas the previous method might hit a few quads before and after. By reducing the quads that are sent to `pcolor` vs. `pcolormesh`, we can save a little run time. Additionally, by not using a relative measure, we don't break when a single-column `pcolormesh` is used, as codified in the test in the last commit.

Some timing shows a small to moderate improvement (averaged over 5 runs):
* Running 14 `pcolor` tests with `pytest -k pcolor`: 7.2±0.2 s to 6.38±0.03 s = 0.8±0.2 s
* Running [this script](https://gist.github.com/QuLogic/747ac39df0ddfe159b49f9e8c792eee6) from #700 with `NROW=NCOL=1`: 2.99±0.03 s to 2.50±0.08 s = 0.49±0.09 s
* Running the same script with `NROW=NCOL=2`: 11.2±0.2 s to 8.94±0.16 s = 2.3±0.3 s
* Running the same script with `NROW=NCOL=10`: 283±4 s to 224±8 s = 59.0±9 s ~ 1 minute!